return the tokenID on assetRegistry's mint function and forge fmt

+++ src/assetNFT.sol
@@ -4,17 +4,16 @@ import "../lib/openzeppelin-contracts/contracts/token/ERC721/ERC721.sol";
 import "tinlake-auth/auth.sol";
 
 contract AssetNFT is ERC721, Auth {
-    
-    uint public count = 0;
+    uint256 public count = 0;
 
     constructor() ERC721("Tinlake Asset NFT", "TNLK") {
         wards[msg.sender] = 1;
         emit Rely(msg.sender);
     }
 
-    function mintTo(address usr) public auth returns(uint) {
+    function mintTo(address usr) public auth returns (uint256) {
         count++;
-        _mint(usr, count); 
+        _mint(usr, count);
         return count;
     }
 }
